fix: some fixes

+++ utils/install_fonts.sh
@@ -5,8 +5,7 @@ set -e
 FONT="JetBrainsMono"
 FONT_PATH="$FONT.tar.xz"
 
-fc-match -a | grep -q $FONT
-if ! [ $? ]; then
+if ! [ $(fc-match -a | grep -q $FONT) ]; then
 	curl -OL https://github.com/ryanoasis/nerd-fonts/releases/latest/download/$FONT_PATH
 	tar -xf ./$FONT_PATH -C ~/.local/share/fonts/fonts/
 	rm ./$FONT_PATH
